Make happy path test pass

+++ src/main/java/com/github/phillbarber/nflow/remoteservices/DiscountPriceResponse.java
@@ -2,5 +2,5 @@ package com.github.phillbarber.nflow.remoteservices;
 
 import java.math.BigDecimal;
 
-public record DiscountPriceResponse (BigDecimal discount, String promotionCode, Integer totalPrice) {
+public record DiscountPriceResponse (BigDecimal discount, String promotionCode, BigDecimal totalPrice) {
 }
